requirements.txt
absl-py==2.2.2
accelerate==1.6.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
Authlib==1.3.1
av==14.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
black==23.12.1
blis==1.3.0
blockbuster==1.5.24
BM25==1.0.0
Bottleneck==1.4.2
cachetools==5.5.2
catalogue==2.0.10
cbor==1.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.2
csvw==3.5.1
ctranslate2==4.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.14
decorator==4.4.2
deprecation==2.1.0
dill==0.3.8
Distance==0.1.3
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
dydantic==0.0.8
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
encodec==0.1.1
espeakng-loader==0.2.4
executing==2.2.0
fastapi==0.115.12
fastembed==0.6.1
faster-whisper==1.1.1
fastrtc==0.0.23
fastrtc-moonshine-onnx==20241016
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
fireworks-ai==0.15.12
FlagEmbedding==1.3.4
flake8==7.2.0
Flake8-pyproject==1.2.3
flatbuffers==25.2.10
fonttools==4.57.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-aiplatform==1.90.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.29.0
gradio_client==1.10.0
greenlet==3.2.1
groovy==0.1.2
groq==0.23.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hf-xet==1.1.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
ijson==3.3.0
imageio==2.25.1
imageio-ffmpeg==0.6.0
inflect==7.5.0
iniconfig==2.1.0
inscriptis==2.6.0
IProgress==0.4
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
ir_datasets==0.5.10
isodate==0.7.2
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
kokoro-onnx==0.4.8
langchain==0.3.24
langchain-anthropic==0.3.12
langchain-community==0.3.22
langchain-core==0.3.56
langchain-deepseek==0.1.3
langchain-fireworks==0.3.0
langchain-google-genai==2.1.3
langchain-google-vertexai==2.0.21
langchain-groq==0.3.2
langchain-openai==0.3.14
langchain-qdrant==0.2.0
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.4.5
langgraph-api==0.1.16
langgraph-checkpoint==2.0.26
langgraph-checkpoint-postgres==2.0.21
langgraph-cli==0.2.7
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.6
langgraph-sdk==0.1.63
langmem==0.0.24
langsmith==0.3.33
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
misaki==0.9.4
mlx==0.24.1
mlx-lm==0.22.3
mlx-vlm==0.1.21
mlx-whisper==0.4.2
mmh3==5.1.0
more-itertools==10.7.0
moviepy==2.0.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.15.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numexpr==2.10.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.21.1
openai==1.76.0
opencv-python==4.10.0.84
orjson==3.10.16
ormsgpack==1.9.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peft==0.15.2
pexels_apis==0.0.6
pexpect==4.9.0
phonemizer==3.3.0
phonemizer-fork==3.3.1
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
preshed==3.0.9
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
pyee==13.0.0
pyflakes==3.3.2
pyfoal==1.0.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pyloudnorm==0.1.1
PyMuPDF==1.25.5
pyOpenSSL==25.0.0
pypar==0.0.6
pyparsing==3.2.3
pypdf==5.4.0
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.14.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
segments==2.3.0
selenium==4.32.0
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==79.0.1
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.2
structlog==25.2.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trec-car-tools==2.6
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.0
trustcall==0.0.39
typeguard==4.4.2
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
unlzw3==0.2.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
validators==0.34.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.14.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yapecs==0.4.0
yarl==1.20.0
zlib-state==0.1.9
zstandard==0.23.0
